Read Desc.
Attempted to build UI with React, Vite, and MantimeUI, but there is a logical problem. Riot Games requires 2FA which severely complicates the problem, as well as the issue of being forced to save the unhashed password to a database, which I am not really comfortable with. The new goal will be to use pywebview with React alongside Flask, or Electron. Will have to decide.