Auth system complete
All the authentications are now complete and finished. Users can login, register, reset password, and all logic is handled on the client side in order to ensure that users do not misbehave. TODO: Add types to Python server, as well as do server-side checks on all emails and passwords. Add password guidelines on client. Set a passphrase register (for now so that the client can be tested and then slowly ramped up). Try using HCaptcha solver to bypass Riot's captcha requirement for API.