Began Riot Auth logic
Using a special redirect link from valapidocs, I am able to get the auth_token and entitlements from Riot in order to be able to check logic. Now, the ET and AT will stay on the client and be sent with each request. I am getting really close to a working prototype, which is really exciting!